refactor: merge into a single pkg
