+++ botatobot/txt2img.go
@@ -1,4 +1,4 @@
-package scheduler
+package botatobot
 
 import (
 	"encoding/base64"
@@ -291,7 +291,6 @@ func localRunner(j *Txt2img) ([]byte, error) {
 }
 
 func remoteRunner(j *Txt2img) ([]byte, error) {
-	fmt.Println("got job")
 	input, err := json.Marshal(j)
 
 	if err != nil {
